+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - (Bugfix) Skip Replace operation on DBServer if they need to be scaled down
 - (Feature) Upgrade procedure steps
 - (Refactor) Remove API and Core cross-dependency
+- (Bugfix) Allow to have nil architecture (NPE fix)
 
 ## [1.2.8](https://github.com/arangodb/kube-arangodb/tree/1.2.8) (2022-02-24)
 - Do not check License V2 on Community images